Move profile details and update user data components
